+++ utils/changelog.md
@@ -7,6 +7,7 @@
 - Change the creation helpers in `Vulkan.Utils.Initialization` to use the new
   `Requirement` types.
 - Add QuasiQuoters for creating device requirements in `Vulkan.Utils.Requirements.TH`
+- Allow selecting target environment in `compileShader`
 
 ## [0.3] - 2020-11-24